removed need for second jupyter port
