+++ jupyter_ascending/extension.py
@@ -1,15 +1,15 @@
-import signal
 import time
 
 from jupyter_ascending._environment import SYNC_EXTENSION
 from jupyter_ascending.handlers import jupyter_notebook
-from jupyter_ascending.handlers import jupyter_server
+import jupyter_ascending.handlers.server_extension
 from jupyter_ascending.logger import J_LOGGER
 from jupyter_ascending.utils import get_name_from_python
 
 
 @J_LOGGER.catch
 def load_ipython_extension(ipython):
+    """This is the specially named function that Jupyter will call to load a notebook extension."""
     set_everything_up()
 
 
@@ -27,30 +27,12 @@ def set_everything_up():
         return
 
     J_LOGGER.info("IPYTHON LOAD: " + time.ctime() + ": " + notebook_name)
-    jupyter_notebook.start_notebook_server_in_thread(notebook_name, jupyter_server)
+    jupyter_notebook.start_notebook_server_in_thread(notebook_name)
 
 
 def load_jupyter_server_extension(ipython):
-    ipython.log.info("LOADING SERVER")
+    """This is the specially named function that Jupyter will call to load a server extension."""
+    ipython.log.info("LOADING JUPYTER ASCENDING SERVER PLUGIN")
     J_LOGGER.info("SERVER LOAD: " + time.ctime())
 
-    server = jupyter_server.start_server_in_thread()
-    if not server:
-        return
-
-    # HACK:
-    # A bit of a hack to make sure the server gets shutdown when we're done here.
-    #   Had some problems with hanging servers
-    #
-    # I think this doesn't quite work if we don't confirm that we want the server shutdown.
-    #   Oh well for now...
-    ORIGINAL = None
-
-    def shutdown_from_signal(*args, **kwargs):
-        if ORIGINAL:
-            ORIGINAL(*args, **kwargs)
-
-        J_LOGGER.info("SERVER: Shutting down server")
-        server.shutdown()
-
-    ORIGINAL = signal.signal(signal.SIGINT, shutdown_from_signal)
+    jupyter_ascending.handlers.server_extension.load_extension(ipython)
