--- jupyter_ascending/handlers/jupyter_server.py
@@ -1,178 +1,0 @@
-"""
-This file contains code for our server that runs alongside the Jupyter server.
-
-It receives commands from the client library in `requests` (potentially from a remote computer),
-and also receives messages from notebooks looking to register to receive Jupyter Ascending
-commands. Commands from the client library are passed along to the appropriate notebook which
-best matches the filepath of the `.sync.py` file.
-"""
-import threading
-from http.server import HTTPServer
-from pathlib import Path
-from typing import Any
-from typing import Dict
-from typing import Optional
-from typing import TypeVar
-
-import attr
-from jsonrpcclient import request
-from jsonrpcclient.exceptions import ReceivedNon2xxResponseError
-from requests.exceptions import ConnectionError
-
-from jupyter_ascending._environment import EXECUTE_HOST_LOCATION
-from jupyter_ascending._environment import EXECUTE_HOST_URL
-from jupyter_ascending._environment import SYNC_EXTENSION
-from jupyter_ascending.errors import UnableToFindNotebookException
-from jupyter_ascending.functional import get_matching_tail_tokens
-from jupyter_ascending.handlers import ServerMethods
-from jupyter_ascending.handlers import generate_request_handler
-from jupyter_ascending.json_requests import JsonBaseRequest
-from jupyter_ascending.logger import J_LOGGER
-
-GenericJsonRequest = TypeVar("GenericJsonRequest", bound=JsonBaseRequest)
-T = TypeVar("T")
-
-_REGISTERED_SERVERS: Dict[str, int] = {}
-
-
-def _clear_registered_servers():
-    global _REGISTERED_SERVERS
-
-    _REGISTERED_SERVERS = {}
-
-
-def _make_url(notebook_port: int):
-    return f"http://localhost:{notebook_port}"
-
-
-# These are the JSON-RPC messages that this server responds to.
-# We'll add more using decorators
-multiplexer_methods = ServerMethods("JupyterServer Start", "JupyterServer Close")
-
-
-@multiplexer_methods.add
-def register_notebook_server(notebook_path: str, port_number: int) -> None:
-    # TODO: why not just inline this?
-    register_server(notebook_path=notebook_path, port_number=port_number)
-
-
-@multiplexer_methods.add
-def perform_notebook_request(notebook_path: str, command_name: str, data: Dict[str, Any]) -> Optional[Dict]:
-    """Receives a command from the client library, picks the notebook that matches
-    the filepath, and forwards the command along to that notebook."""
-    J_LOGGER.debug("Performing notebook request... ")
-
-    try:
-        notebook_server = get_server_for_notebook(notebook_path)
-    except UnableToFindNotebookException:
-        J_LOGGER.warning(f"Unabled to find {notebook_path} in {_REGISTERED_SERVERS}")
-        return {"success": False, "notebook_path": notebook_path}
-
-    request(notebook_server, command_name, data=data)
-
-    return None
-
-
-# This is the HTTP POST handler that dispatches the JSON-RPC messages to the appropriate function
-JupyterServerRequestHandler = generate_request_handler("JupyterServer", multiplexer_methods)
-
-
-def register_server(notebook_path: str, port_number: int) -> None:
-    """Receives a message from a notebook requesting to register itself
-    as available to receive Jupyter Ascending commands."""
-    J_LOGGER.info("Registering notebook {notebook} on port {port}", notebook=notebook_path, port=port_number)
-
-    _REGISTERED_SERVERS[notebook_path] = port_number
-
-    J_LOGGER.debug("Updated notebook mappings: {}", _REGISTERED_SERVERS)
-
-
-def get_server_for_notebook(notebook_str: str) -> Optional[str]:
-    """Get the URL to the server running on the Jupyter notebook that best matches this filename."""
-    # Normalize to notebook path
-    notebook_str = notebook_str.replace(f".{SYNC_EXTENSION}.py", f".{SYNC_EXTENSION}.ipynb")
-    J_LOGGER.debug("Finding server for notebook_str, script_path: {}", notebook_str)
-
-    notebook_path = Path(notebook_str)
-
-    def get_score_for_name(registered_name: str) -> int:
-        """
-        Note that it is matching on parts of a path
-
-        Returns the consecutive count of matching parts of a path, from the end toward the start.
-
-        registered ['tmp', 'notebooks', 'myfile.py']
-        notebook   ['opt', 'notebooks', 'myfile.py']
-         -> 2
-
-        registered ['a', 'b', 'c']
-        notebook   ['a', 'b', 'd']
-         -> 0
-
-        """
-        return len(get_matching_tail_tokens(notebook_path.parts, Path(registered_name).parts))
-
-    score_by_name = {x: get_score_for_name(x) for x in _REGISTERED_SERVERS.keys()}
-
-    if len(score_by_name) == 0:
-        raise UnableToFindNotebookException(f"No registered notebooks")
-
-    max_score = max(score_by_name.values())
-
-    if max_score <= 0:
-        raise UnableToFindNotebookException(f"Could not find server for notebook_str: {notebook_str}")
-
-    # Only found one reasonable notebook.
-    best_scores = [k for k, v in score_by_name.items() if v == max_score]
-
-    if len(best_scores) == 1:
-        notebook_port = _REGISTERED_SERVERS[best_scores[0]]
-
-        J_LOGGER.debug("Found server at port {}", notebook_port)
-        return _make_url(notebook_port)
-    else:
-        raise UnableToFindNotebookException(f"Could not find server for notebook_str: {notebook_str}")
-
-
-def request_notebook_command(json_request: GenericJsonRequest):
-    """This is a command to be used by the client libraries to send a command to this server.
-
-    It calls unpacks the JsonRequest and calls `perform_notebook_request` defined above."""
-    try:
-        result = request(
-            EXECUTE_HOST_URL,
-            perform_notebook_request.__name__,
-            command_name=type(json_request).__name__,
-            notebook_path=json_request.file_name,
-            data=attr.asdict(json_request),
-        )
-
-        if not result.data.result:
-            return
-
-        if not result.data.result.get("success", True):
-            raise Exception(f"Failed to complete request. {result.data}")
-
-    except ConnectionError as e:
-        J_LOGGER.error(f"Unable to connect to server. Perhaps notebook is not running? {e}")
-    except ReceivedNon2xxResponseError as e:
-        J_LOGGER.error(f"Unable to process request. Perhaps something else is running on this port? {e}")
-
-
-def start_server_in_thread():
-    """Run a server handling command requests from client libraries and registration requests from
-    notebooks."""
-    try:
-        server_executor = HTTPServer(EXECUTE_HOST_LOCATION, JupyterServerRequestHandler)
-    except OSError:
-        print(f"It appears you already are using {EXECUTE_HOST_LOCATION}")
-        print("Use the environment variable: 'JUPYTER_ASCENDING_EXECUTE_PORT' to set a new port")
-
-        return
-
-    server_executor_thread = threading.Thread(target=server_executor.serve_forever, args=tuple())
-    server_executor_thread.start()
-
-    J_LOGGER.info("Successfully started multiplexer server")
-
-    return server_executor
