+++ jupyter_ascending/requests/execute_all.py
@@ -1,9 +1,9 @@
 import argparse
 from pathlib import Path
 
-from jupyter_ascending.handlers import jupyter_server
 from jupyter_ascending.json_requests import ExecuteAllRequest
 from jupyter_ascending.logger import J_LOGGER
+from jupyter_ascending.requests.client_lib import request_notebook_command
 
 
 def send(file_name: str):
@@ -11,7 +11,7 @@ def send(file_name: str):
     file_name = str(Path(file_name).absolute())
 
     request_obj = ExecuteAllRequest(file_name=file_name)
-    jupyter_server.request_notebook_command(request_obj)
+    request_notebook_command(request_obj)
 
     J_LOGGER.info("... Complete")
