+++ jupyter_ascending/requests/get_status.py
@@ -1,15 +1,15 @@
 import argparse
 from pathlib import Path
 
-from jupyter_ascending.handlers import jupyter_server
 from jupyter_ascending.json_requests import GetStatusRequest
 from jupyter_ascending.logger import J_LOGGER
+from jupyter_ascending.requests.client_lib import request_notebook_command
 
 def send(file_name: str):
     file_name = str(Path(file_name).absolute())
 
     request_obj = GetStatusRequest(file_name=file_name)
-    jupyter_server.request_notebook_command(request_obj)
+    request_notebook_command(request_obj)
 
 
 if __name__ == "__main__":
